+++ frontend/src/js/commands/hotPixel.js
@@ -17,7 +17,7 @@ export default params => {
     const viewerID = params.viewer_id;
     const viewer = store.getState().viewers[viewerID];
     const region = params.region;
-    const threshold = (params.threshold === "max") ? "max" : parseInt(params.threshold);
+    const threshold = (params.threshold === "max") ? "max" : parseFloat(params.threshold);
 
     // The parameters to pass to the backend
     const backendParameters = {
